[refactor] model 파일명 수정 - 복수형, 단수형 통일을 위해 model 파일명 수정함
